refactor(connect): move connectService and jolokiaService to shared #435
Fix #435
