+++ packages/hawtio/src/plugins/camel/exchanges/BlockedExchanges.test.tsx
@@ -1,13 +1,12 @@
-import { render, screen, fireEvent, waitForElementToBeRemoved } from '@testing-library/react'
-import { MBeanNode, MBeanTree, workspace } from '@hawtiosrc/plugins/shared'
-import { CamelContext } from '../context'
-import { BlockedExchanges } from './BlockedExchanges'
-import { jmxDomain } from '../globals'
 import { camelTreeProcessor } from '@hawtiosrc/plugins/camel/tree-processor'
-import { Exchange } from './exchanges-service'
-import { jolokiaService } from '@hawtiosrc/plugins/connect'
+import { MBeanNode, MBeanTree, jolokiaService, workspace } from '@hawtiosrc/plugins/shared'
+import { fireEvent, render, screen, waitForElementToBeRemoved } from '@testing-library/react'
 import fs from 'fs'
 import path from 'path'
+import { CamelContext } from '../context'
+import { jmxDomain } from '../globals'
+import { BlockedExchanges } from './BlockedExchanges'
+import { Exchange } from './exchanges-service'
 
 const routesXmlPath = path.resolve(__dirname, '..', 'testdata', 'camel-sample-app-routes.xml')
 const sampleRoutesXml = fs.readFileSync(routesXmlPath, { encoding: 'utf8', flag: 'r' })
@@ -15,7 +14,7 @@ const sampleRoutesXml = fs.readFileSync(routesXmlPath, { encoding: 'utf8', flag:
 /**
  * Mock the routes xml to provide a full tree
  */
-jest.mock('@hawtiosrc/plugins/connect/jolokia-service')
+jest.mock('@hawtiosrc/plugins/shared/jolokia-service')
 jolokiaService.execute = jest.fn(async (mbean: string, operation: string, args?: unknown[]): Promise<unknown> => {
   if (
     mbean === 'org.apache.camel:context=SampleCamel,type=context,name="SampleCamel"' &&
