+++ packages/hawtio/src/plugins/camel/type-converters/type-converters-service.test.tsx
@@ -1,11 +1,9 @@
-import { MBeanNode, MBeanTree, workspace } from '@hawtiosrc/plugins/shared'
-import { jmxDomain } from '../globals'
 import { camelTreeProcessor } from '@hawtiosrc/plugins/camel/tree-processor'
-import { jolokiaService } from '@hawtiosrc/plugins/connect'
-import * as tcs from './type-converters-service'
+import { AttributeValues, MBeanNode, MBeanTree, jolokiaService, workspace } from '@hawtiosrc/plugins/shared'
 import fs from 'fs'
 import path from 'path'
-import { AttributeValues } from '@hawtiosrc/plugins/connect/jolokia-service'
+import { jmxDomain } from '../globals'
+import * as tcs from './type-converters-service'
 
 const routesXmlPath = path.resolve(__dirname, '..', 'testdata', 'camel-sample-app-routes.xml')
 const sampleRoutesXml = fs.readFileSync(routesXmlPath, { encoding: 'utf8', flag: 'r' })
@@ -22,7 +20,7 @@ let canDisplayTypeConvertersStatistics = false
 /**
  * Mock the routes xml to provide a full tree
  */
-jest.mock('@hawtiosrc/plugins/connect/jolokia-service')
+jest.mock('@hawtiosrc/plugins/shared/jolokia-service')
 jolokiaService.execute = jest.fn(async (mbean: string, operation: string, args?: unknown[]): Promise<unknown> => {
   if (
     mbean === 'org.apache.camel:context=SampleCamel,type=context,name="SampleCamel"' &&
