+++ packages/hawtio/src/plugins/camel/type-converters/type-converters-service.ts
@@ -1,4 +1,4 @@
-import { jolokiaService } from '@hawtiosrc/plugins/connect'
+import { jolokiaService } from '@hawtiosrc/plugins/shared'
 import { MBeanNode } from '@hawtiosrc/plugins/shared/tree'
 import { findContext } from '../camel-service'
 import { mbeansType } from '../globals'
