+++ packages/hawtio/src/plugins/shared/__mocks__/jmx-camel-tree.json
