+++ packages/hawtio/src/plugins/shared/__mocks__/jolokia-service.ts
