+++ packages/hawtio/src/plugins/shared/chart/Chart.tsx
@@ -1,5 +1,5 @@
-import { AttributeValues } from '@hawtiosrc/plugins/connect/jolokia-service'
 import { PluginNodeSelectionContext } from '@hawtiosrc/plugins/context'
+import { AttributeValues } from '@hawtiosrc/plugins/shared/jolokia-service'
 import { MBeanNode } from '@hawtiosrc/plugins/shared/tree'
 import { isNumber } from '@hawtiosrc/util/objects'
 import { ChartArea, ChartAxis, Chart as ChartDraw, ChartVoronoiContainer } from '@patternfly/react-charts'
