+++ packages/hawtio/src/plugins/shared/index.ts
@@ -4,6 +4,8 @@ export { JmxContentMBeans } from './JmxContentMBeans'
 export { PluginTreeViewToolbar } from './PluginTreeViewToolbar'
 export { AttributeModal, Attributes } from './attributes'
 export { Chart } from './chart'
+export * from './connect-service'
+export * from './jolokia-service'
 export { Operations } from './operations'
 export * from './tree'
 export * from './workspace'
