+++ packages/hawtio/src/plugins/shared/tree/node.test.ts
@@ -3,7 +3,7 @@ import { workspace } from '../workspace'
 import { Icons, MBeanNode, PropertyList } from './node'
 import { MBeanTree } from './tree'
 
-jest.mock('@hawtiosrc/plugins/connect/jolokia-service')
+jest.mock('@hawtiosrc/plugins/shared/jolokia-service')
 
 describe('MBeanNode', () => {
   let tree: MBeanTree
